ep: use CUDA atomic-compatible NCCL barrier epoch type

+++ mooncake-ep/include/elastic/mooncake_ep_elastic_comm.cuh
@@ -111,7 +111,7 @@ __forceinline__ __device__ void mooncake_barrier_wo_local_sync(
         // Each tag owns a monotonic receive counter in the symmetric workspace.
         // LSA peers update it with a system atomic and remote peers use a GIN
         // VA signal-add, so no CUDA IPC pointer or IBGDA atomic is involved.
-        auto* epoch = reinterpret_cast<uint64_t*>(
+        auto* epoch = reinterpret_cast<unsigned long long*>(
             workspace.get_nvl_barrier_counter_ptr(kTag));
         auto* arrivals = reinterpret_cast<uint64_t*>(
             workspace.get_nvl_barrier_signal_ptr(kTag, 0));
@@ -122,10 +122,10 @@ __forceinline__ __device__ void mooncake_barrier_wo_local_sync(
                                        0);
         }
         __syncthreads();
-        __shared__ uint64_t target;
+        __shared__ unsigned long long target;
         if (thread_idx == 0)
             target = (atomicAdd(epoch, 1ULL) + 1ULL) *
-                     static_cast<uint64_t>(kNumRanks);
+                     static_cast<unsigned long long>(kNumRanks);
         __syncthreads();
         timeout_while<kNumTimeoutCycles>(
             thread_idx == 0, [=](const bool& is_last_check) {
